requirements.txt
pyyaml
numpy
tqdm
tiktoken